1. добавлен метод для загрузки компании с заблокированными сущностями 2. убраны из модели FETCH.JOIN
