+++ src/main/java/spring/app/dao/abstraction/CompanyDao.java
@@ -2,13 +2,14 @@ package spring.app.dao.abstraction;
 
 import spring.app.model.Company;
 
+import java.util.List;
+
 public interface CompanyDao extends GenericDao<Long, Company> {
     Company getCompanyByCompanyName(String companyName);
 
     /**
      * Получает компанию с заблокированными сущнстями Genre/Song/Author
-     * @param id
      * @return Company
      */
-    //Company getCompanyWithEntityBanned(Long id);
+    Company getCompanyWithEntityBanned(long id);
 }
